requirements.txt
requests
beautifulsoup4
Pillow
pillow-heif
